Implemented new function m/volume.m to reduce repetition
Many files in the project had the same 3 lines of code used to calculate a
volume from a given set of nodes, tetrahedra and an index matrix. These were
extracted into a function to reduce repetition. The funtion was also taken
into use in the culprit files.

A bunch of leftover whitespace was also removed from the related files. Matlab
should really do this automatically…