A variable name fix + an update to volume function
There were some typos left over at the call sites of the new function
m/volume.m, which should be resolved in this commit. They would have resulted
in errors related to unknown variables.

The definition of the index matrix ind_m was also moved inside of the volume
function, as the matrix was constant across all of the function's invocations.

The file m/forward_scripts/brain/lead_field_eeg_fem.m was indented properly,
to make it easier to read. This will show up in the commit diff as if the
entire file had changed. Its semantic contents should not differ, apart from
the new invocation of the volume function with one less argument.