Removed trailing whitespace and multiple blank lines from all text files
A cleanup commit. This was achieved with the command

	find . -not \(                  \
		-name .svn -prune           \
		-o -name .git -prune        \
		-o -name "*.png" -prune     \
		-o -name "*.jpg" -prune     \
		-o -name "*.fig" -prune     \
		-o -name "*.stl" -prune     \
		-o -name "*.mlapp" -prune   \
		-o -name "*.avi" -prune     \
		-o -name "*.mat" -prune     \
	\)                              \
	-type f                         \
	-exec sed -i "s/[[:space:]]\+$//g" "{}" \;

executed in project root. The folders and file types which were ignored can be
found in the "-not \(…\)" list above. The folders .git and .svn were
especially important to ignore, as modifying them would mess up project
history.

Also collapsed all multiple blank lines into one in all text files with a
similar command as above, but with the first argument to `sed` changed to

	'N;/^\n$/D;P;D;'

instead of

	"s/[[:space:]]\+$//g"

Too much empty space everywhere…