zef_tetra_volume now takes a 3rd parameter take_absolute_value
The absolute value of the volume is returned if this is set. Also cleaned up
line breaks so what is added to what is clearer.

The volume calculations that used to have an absolute value related with them
in commit 47d4d4e , as in

        tilavuus = abs(…);

have take_absolute_value set to true. It seems like this was pretty much all
of the already modified ones. Maybe the error is not related to the volume
calculation either...?

Also changed calls to volume/2 → zef_tetra_volume/3 under plugins/. These were
either not updated properly in a previous commit, or were changed back as the
files were restored from history. Either way, the plugins that need it now use
the new volume calculation.