Added a new function m/zef_build_elecrodes.m
The function constructs the matrices B and C from a given stiffness matrix A
and electrode indices. The appropriate sizes for B and C need to be provided
by the calling context as well.

The function will be introduced into the lead field matrix calculations in
m/forward_scripts/brain/lead_field_*_fem.m later, to make the code in there
easier to read.