lead_field_tes_fem: moved PEM code into zef_build_electrodes
The bit of code that did the building of B and C in the PEM case was moved
inside of zef_build_electrodes as well, in addition to the original CEM case.
The function zef_build_electrode chooses the code to run based on the given
elecrode model, and informs the user in the console if the given electrode
model was unknown.