lead_field_tes_fem: stiffness matrix A now returned from zef_transfer_matrix
The function modifies A, but did not return the modified copy in prior
versions. If this turns out to be significant from the point of view of L_tes,
it might be necessary to define a new handle class MatrixCarrier < handle,
that allows forcing the passage of matrices into functions as pointers, even
if they are modified inside of the function.

See https://se.mathworks.com/matlabcentral/answers/109304-passing-matrix-using-call-by-reference#answer_117944
for details on how to create and use such a handle class.