lead_field_eeg_fem: replaced electrode matrix calculation code with a call to zef_build_electrodes
The results differ slightly from the prior EEG electrode construction
implementation still present in commit d75935d: maximum elementwise absolute
difference in zef.L after and before this change is ~64. The calculation is
now performed identically with the tES routine, however, where the electrode
matrices' 𝐵 and 𝐶 construction algorithm was different from the EEG routine to
begin with, possibly unintentionally based on today's discussions, as it was
written later than the EEG case.