lead_field_eeg_fem: added functions zef_{fi,ew}_dipoles
The functions are used to calculate the vectors needed in forming the stensils
or adjacency matrices T_{fi,ew} and divergence matrices G_{fi,ew} in both the
face-intersecting and edgewise cases. This shortens the EEG routine by quite a
bit, but does not seem to have an effect on the calculated lead field L_eeg.