fs2zef: initial parametrization and fixing of input and output paths
All parameters (subjects directory, individual subject folder in the subjects
directory and FreeSurfer (FS) output directory) are read as inputs to the
script. This means that the script has to be called as

	fs2zef.sh <subjects dir> <individual subject dir> <FS output dit>

The existence of the required directories are checked and if they do not
exist, the script does nothing but prints an error message to the screen.