Add a .gitattributes file
This file tells Git to change the line endings in newly added text files to a
simple Line Feed <LF> instead of the Windows equivalent <CR><LF>, as the files
are added to the working tree (the set of files that Git is keeping track of)
in this repository. This makes the reading experience of Unix users a bit more
pleasant. The line feeds are converted back to <CR><LF> on Windows as they are
checked out for local editing. Binary files such as mlapp are ignored and not
modified.

---------------------------------

Allowed .gitattributes

Added an initial .gitattributes file

Explicitly declare line ending type for text files

Added text attribute to README

Added the binary attribute to .fig files

Added the text attribute to .asc files

Added the binary attribute to .mat and .stl files

Fixed typo in comment and added a spaces between file spec and binary attributes