zef_process_inversion: remove many unnecessary if-blocks
These can be added back later if they are deemed necessary, but for now
considering different numbers of inputs and outputs and worrying about the
contents of the input inverter are unnecessary. The structure of the inverter
class is partly checked by the class itself, and the isAnInverter validator
method.

Also simplified how the waitbars are updated.

NOTE: it seems that z_vec was invisible to Matlab in the previous update, when
it was being read into z_inverse. It is weird that sometimes Matlab decides it
is lexically scoped and sometimes it is not.